Remove unneeded condition.
For `follow`, let laravel do this for you. `$user->followers()->sync($this->id, false);` works too.

For `unFollow`, remove the conditional and it will still work fine. If you need to know if anything was deleted, `detach` uses `delete` under the hood and returns the number of rows affected.

I would also rename `unFollow` to `unfollow`.